Move installed repository loading to a trait
